Struct ariadne::Label

pub struct Label<S = Range<usize>> { /* private fields */ }

A type that represents a labelled section of source code.

Implementations

impl<S: Span> Label<S>

pub fn new(span: S) -> Self

Create a new Label. If the span is specified as a Range<usize> the numbers have to be zero-indexed character offsets.

Panics

Panics if the given span is backwards.

pub fn with_message<M: ToString>(self, msg: M) -> Self

Give this label a message.

pub fn with_color(self, color: Color) -> Self

Give this label a highlight colour.

pub fn with_order(self, order: i32) -> Self

Specify the order of this label relative to other labels.

Lower values correspond to this label having an earlier order.

If unspecified, labels default to an order of 0.

When labels are displayed after a line the crate needs to decide which labels should be displayed first. By Default, the orders labels based on where their associated line meets the text (see LabelAttach). Additionally, multi-line labels are ordered before inline labels. You can use this function to override this behaviour.

pub fn with_priority(self, priority: i32) -> Self

Specify the priority of this label relative to other labels.

Higher values correspond to this label having a higher priority.

If unspecified, labels default to a priority of 0.

Label spans can overlap. When this happens, the crate needs to decide which labels to prioritise for various purposes such as highlighting. By default, spans with a smaller length get a higher priority. You can use this function to override this behaviour.