ariadne

Trait Fmt

pub trait Fmt: Sized {
    // Provided methods
    fn fg<C: Into<Option<Color>>>(self, color: C) -> Foreground<Self>
       where Self: Display { ... }
    fn bg<C: Into<Option<Color>>>(self, color: C) -> Background<Self>
       where Self: Display { ... }
}

A trait used to add formatting attributes to displayable items.

If using the concolor feature, this trait assumes that the items are going to be printed to stderr. If you are printing to stdout, use the StdoutFmt trait instead.

Attributes specified through this trait are not composable (i.e: the behaviour of two nested attributes each with a conflicting attribute is left unspecified).

Provided Methods

fn fg<C: Into<Option<Color>>>(self, color: C) -> Foreground<Self>
where Self: Display,

Give this value the specified foreground colour.

fn bg<C: Into<Option<Color>>>(self, color: C) -> Background<Self>
where Self: Display,

Give this value the specified background colour.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.