ariadne

Trait Cache

pub trait Cache<Id: ?Sized> {
    type Storage: AsRef<str>;

    // Required methods
    fn fetch(&mut self, id: &Id) -> Result<&Source<Self::Storage>, impl Debug>;
    fn display<'a>(&self, id: &'a Id) -> Option<impl Display + 'a>;
}

A trait implemented by Source caches.

Required Associated Types

type Storage: AsRef<str>

The type used to store the string data for this cache.

Alternative types other than String can be used, but at the moment, the storage must be contiguous. A primary use case for this is to use a reference-counted string instead of copying the whole contents into a Source.

Required Methods

fn fetch(&mut self, id: &Id) -> Result<&Source<Self::Storage>, impl Debug>

Fetch the Source identified by the given ID, if possible.

fn display<'a>(&self, id: &'a Id) -> Option<impl Display + 'a>

Display the given ID. as a single inline value.

This function may make use of attributes from the Fmt trait.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.